{"schema":"decodex/commit/1","summary":"Extract capture host live preview","authority":"manual"}
