+++ docs/reference/host-core-reset.md
@@ -140,6 +140,8 @@ The current native-host Swift split is:
   with per-track throttling state, and AppKit-to-controller pointer delivery support.
 - `CaptureHostView+LivePrimaryInteraction.swift`: capture-host live primary interaction release
   recovery, pointer-preview mutation, mouse-up monitor wiring, and release-watchdog orchestration.
+- `CaptureHostView+LivePreview.swift`: capture-host live preview snapshots, HUD/loupe placement,
+  live sample-cache use, and controller preview-demand updates.
 - `LiveOverlayRenderer.swift`: live overlay render orchestration for HUD, loupe, frozen-pending
   preview, and live chrome layers.
 - `LiveHudColorRollCoordinator.swift`: live HUD color swatch/hex presentation, pending color roll
